fix orderId and openId init

+++ src/app/index.run.js
@@ -14,25 +14,31 @@
     });
 
     // orderId & openId init
-    var search = utils.getLocationSearch(),
-        $search = $location.search(),
-        order = localStorageService.get('order');
+    var orderInit = function() {
+      var search = utils.getLocationSearch(),
+          $search = $location.search(),
+          order = localStorageService.get('order');
 
-    var _openId = search.openId || $search.openId || Math.floor(Math.random()*1000000),
-        _orderId = search.orderId || $search.orderId || Math.floor(Math.random()*1000000);
+      var _openId = search.openId || $search.openId || Math.floor(Math.random()*1000000),
+          _orderId = search.orderId || $search.orderId || Math.floor(Math.random()*1000000);
 
-    if(!order || (order.openId !== _openId)) {
-      order = {
-        openId: _openId,
-        orderId: _orderId
-      };
+      if(!search.openId && !$search.openId) return;
 
-      localStorageService.remove('user', 'mId', 'sessionId');
-    } else if(order.openId === _openId && order.orderId !== _orderId) {
-      order.orderId = _orderId;
-    }
+      if(!order || (order.openId !== _openId)) {
+        order = {
+          openId: _openId,
+          orderId: _orderId
+        };
 
-    localStorageService.add('order', order);
+        localStorageService.remove('user', 'mId', 'sessionId');
+      } else if(order.openId === _openId && order.orderId !== _orderId) {
+        order.orderId = _orderId;
+      }
+
+      localStorageService.add('order', order);
+    };
+
+    orderInit();
 
     // auto login
     userService.autoLogin();
